catalogo
- catalogo agora carrega a partir do bd
- carrinho parou de funcionar

+++ src/DB/database.sql
@@ -1,10 +1,20 @@
+USE MASTER
+go
+
+IF EXISTS(SELECT 1 FROM SYSDATABASES WHERE NAME = 'anyges')
+	DROP DATABASE anyges
+go
+
 CREATE DATABASE anyges;
+go
 
 USE anyges;
+go
 
-DROP DATABASE anyges
+-- DROP DATABASE anyges
 
 SET DATEFORMAT DMY;
+go
 
 -- Associacao
 
@@ -19,8 +29,8 @@ CREATE TABLE tblAssociacao (
 	cidade_associacao varchar(25) not null,
 	estado_associacao char(2) not null
 );
-
 CREATE INDEX xAssociacao ON tblAssociacao(ID_associacao);
+go
 
 -- Cupom
 
@@ -28,13 +38,14 @@ CREATE TABLE tblCupom (
 	ID_cupom numeric(6) not null PRIMARY KEY identity(1,1),
 	nome_cupom varchar(25) not null,
 	status char(1) not null,
-	valor decimal (8,2) not null,
+	valor numeric (8) not null,
 	tipo varchar (25) not null,
-	imagem varchar (25) not null,
-	descricao_cupom varchar(25) not null
+	imagem varchar (50) not null,
+	descricao_cupom varchar(75) not null,
+	desconto numeric(2) not null
 );
-
 CREATE INDEX xCupom ON tblCupom(ID_cupom);
+go
 
 -- Local Deposito
 
@@ -47,8 +58,8 @@ CREATE TABLE tblLocalDeposito (
 	cidade_deposito varchar(25) not null,
 	estado_deposito char(2) not null
 );
-
 CREATE INDEX xLocalDeposito ON tblLocalDeposito(ID_local_deposito);
+go
 
 -- Usuário
 
@@ -62,8 +73,8 @@ CREATE TABLE tblUsuario (
 	senha_usuario varchar(15) not null,
 	dt_nascimento date not null
 );
-
 CREATE INDEX xUsuario ON tblUsuario(ID_usuario);
+go
 
 -- Feedback
 
@@ -72,8 +83,8 @@ CREATE TABLE tblFeedback (
 	ID_usuario numeric(6) FOREIGN KEY REFERENCES tblUsuario(ID_usuario) not null,
 	descricao_feedback varchar(250) not null
 );
-
 CREATE INDEX xFeedback ON tblFeedback(ID_feedback);
+go
 
 -- Doacao
 
@@ -85,8 +96,8 @@ CREATE TABLE tblDoacao (
 	pontuacao decimal(7,2) not null,
 	descricao_doacao varchar(75)
 );
-
 CREATE INDEX xDoacao ON tblDoacao(ID_doacao);
+go
 
 -- Pedido
 
@@ -95,8 +106,8 @@ CREATE TABLE tblPedido (
 	ID_usuario numeric(6) FOREIGN KEY REFERENCES tblUsuario(ID_usuario) not null,
 	dt_pedido datetime not null
 );
-
 CREATE INDEX xPedido ON tblPedido(ID_pedido);
+go
 
 -- Resgate
 
@@ -106,8 +117,8 @@ CREATE TABLE tblResgate (
 	ID_pedido numeric(6) FOREIGN KEY REFERENCES tblPedido(ID_pedido) not null,
 	qtde_resgate int not null
 );
-
 CREATE INDEX xResgate ON tblResgate(ID_resgate);
+go
 
 -- Associacao cupom
 
@@ -117,9 +128,21 @@ CREATE TABLE tblAssociacaoCupom (
 	ID_cupom numeric(6) FOREIGN KEY REFERENCES tblCupom(ID_cupom) not null,
 	qtde_estoque int
 );
-
 CREATE INDEX xAssociaoCupom ON tblAssociacaoCupom(ID_associacao_cupom);
+go
 
 -- COMANDOS --
 
+INSERT INTO tblCupom VALUES ('Rivotril', 'S', 12000, 'Promofarma', './src/imgs/cupons/rivotril.png', 'Suspensão Oral', 25);
+INSERT INTO tblCupom VALUES ('Pediatra', 'S', 25000, 'Unimed', './src/imgs/cupons/unimed.png', 'Terça-feira e Quinta-feira', 25);
+INSERT INTO tblCupom VALUES ('Psicóloga', 'S', 27000, 'Hospital das Clínicas', './src/imgs/cupons/hc.png', 'Domingo à Quarta-feira', 25);
+INSERT INTO tblCupom VALUES ('Paracetamol 750mg', 'S', 12000, 'Promofarma', './src/imgs/cupons/paracetamol.jpg', 'Comprimido revestido Dor e Febre', 25);
+INSERT INTO tblCupom VALUES ('Clínico Geral', 'S', 12000, 'Amil', './src/imgs/cupons/amil.jpg', 'Terça-feira e Sexta-feira', 25);
+INSERT INTO tblCupom VALUES ('Gastroenterologista', 'S', 12000, 'Sírio Libanês', './src/imgs/cupons/sirio.jpg', 'Quarta-feira à Sexta-feira', 25);
+go
+
+SELECT * FROM tblCupom;
 SELECT * FROM tblUsuario;
+
+SELECT * FROM tblCupom WHERE nome_cupom='Rivotril' AND valor=12000
+go
